Add back button to login vs and rename Controllers -> Scenes fix bug with bottom border navigation vc
