+++ VenueWatch/Scenes/Friends/FriendsViewController.swift
