+++ VenueWatch/Scenes/Navigation/NavigationController.swift
@@ -18,3 +18,17 @@ final class NavigationController: UINavigationController {
         ]
     }
 }
+
+extension UINavigationController {
+    func presentLoginViewController(for type: LoginType) {
+        let loginViewController = LoginViewController(currentLoginType: type)
+        UIView.transition(
+            with: view, duration: 0.5,
+            options: .transitionFlipFromRight,
+            animations: {
+                self.viewControllers.removeLast()
+                self.viewControllers.append(loginViewController)
+        },
+            completion: nil)
+    }
+}
